requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
outcome==1.2.0
pushbullet.py==0.12.0
PySocks==1.7.1
python-magic==0.4.27
pytz==2023.3
requests==2.31.0
schedule==1.2.0
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
websocket-client==1.6.1
wsproto==1.2.0
